+++ crates/miden-lib/src/errors/mod.rs
@@ -9,8 +9,8 @@ pub mod note_script_errors;
 mod masm_error;
 pub use masm_error::MasmError;
 
-mod script_builder_errors;
-pub use script_builder_errors::ScriptBuilderError;
+mod protocol_assembler_errors;
+pub use protocol_assembler_errors::ProtocolAssemblerError;
 
 mod transaction_errors;
 pub use transaction_errors::{TransactionEventError, TransactionTraceParsingError};