+++ crates/miden-lib/src/utils/mod.rs
@@ -1,6 +1,6 @@
-pub mod script_builder;
+pub mod protocol_assembler;
 
 pub use miden_objects::utils::*;
-pub use script_builder::ScriptBuilder;
+pub use protocol_assembler::ProtocolAssembler;
 
-pub use crate::errors::ScriptBuilderError;
+pub use crate::errors::ProtocolAssemblerError;