+++ crates/miden-objects/src/account/component/template/mod.rs
@@ -245,6 +245,9 @@ impl AccountComponentMetadata {
     }
 }
 
+// CONVERSIONS
+// ================================================================================================
+
 impl TryFrom<&Package> for AccountComponentMetadata {
     type Error = AccountError;