+++ crates/miden-testing/src/tx_context/context.rs
@@ -3,9 +3,11 @@ use alloc::collections::{BTreeMap, BTreeSet};
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 
+use miden_lib::testing::mock_account_code::MockAccountCodeExt;
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::account::{
     Account,
+    AccountCode,
     AccountId,
     PartialAccount,
     StorageMapWitness,
@@ -129,6 +131,7 @@ impl TransactionContext {
 
         let assembler = TransactionKernel::with_mock_libraries(self.source_manager.clone())
             .with_debug_mode(true);
+
         let program = assembler
             .with_debug_mode(true)
             .assemble_program(virtual_source_file)
@@ -138,6 +141,11 @@ impl TransactionContext {
         // Note that native and foreign account's code are already loaded by the
         // TransactionContextBuilder.
         self.mast_store.insert(TransactionKernel::library().mast_forest().clone());
+        self.mast_store.insert(miden_lib::StdLibrary::default().mast_forest().clone());
+        self.mast_store.insert(miden_lib::MidenLib::default().mast_forest().clone());
+        self.mast_store
+            .insert(AccountCode::mock_account_library().mast_forest().clone());
+        self.mast_store.insert(AccountCode::mock_faucet_library().mast_forest().clone());
         self.mast_store.insert(program.mast_forest().clone());
 
         let account_procedure_idx_map = AccountProcedureIndexMap::new(