+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Split tx progress events out into a separate enum ([#2103](https://github.com/0xMiden/miden-base/pull/2103)).
 - Added `note::get_network_account_tag` procedure ([#2120](https://github.com/0xMiden/miden-base/pull/2120)).
 - [BREAKING] Removed `AccountComponentTemplate` in favor of instantiating components via `AccountComponent::from_package` ([#2127](https://github.com/0xMiden/miden-base/pull/2127)).
+- [BREAKING] Renamed `ScriptBuilder` into `ProtocolAssembler` and added the ability to build component code ([#2142](https://github.com/0xMiden/miden-base/pull/2142)).
 
 ## 0.12.4 (2025-11-26)